feat: use database and layer hashes
The digests themselves may be different than what we have locally for
one reason or another. Use the layer hashes since that's the important
part.

We also don't really care what we're running right now, we really just
want to know if things upstream have changed. Remove the local query
and compare functions.

Add a database for managing remote layer hashes. It's rough, but it
works for now. Add a flag to clear the database.

Signed-off-by: Randolph Sapp <rs@ti.com>